requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
httpx>=0.27
python-multipart>=0.0.9
python-dotenv>=1.0
